requirements.txt
# ========================================
# Jarvis Commander - Dépendances Python
# ========================================

# Interface graphique
PySide6>=6.5.0
PySide6-Addons>=6.5.0

# Audio de base
sounddevice>=0.4.6
numpy>=1.24.0

# Wake Word Detection (Picovoice Porcupine)
pvporcupine>=3.0.0

# Speech-to-Text (Whisper)
faster-whisper>=0.9.0
# Pour CUDA (optionnel si GPU NVIDIA disponible)
# torch>=2.0.0  # Décommentez si vous utilisez GPU

# Text-to-Speech
pyttsx3>=2.90

# Contrôle système
psutil>=5.9.0
pyautogui>=0.9.54

# Configuration
PyYAML>=6.0

# Utilitaires
pathlib>=1.0.1
